fastapi_app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
openai-whisper==20231117
torch==2.1.1
numpy==1.24.3 